tests/krate/publish: Extract `git` module
